Rename underlying view return version modules
